Added RequestMsgInbox method to the NATS connection.
This new method means you set the Reply subject in a new message which is acknowledged by the client.
A new request is made toward the chosen subject and replies are accepted by publishing or responding.
If responders are not listening, then you will see ErrNoResponders.
It also deals with timeouts. The call can be made in a GR too.

Signed-off-by: davedotdev <me@dave.dev>